float(10) is converted to real, not supported in gus, use float
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonData/trunk@42260 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Load/lib/sql/apidbschema-pg/createPlasmoMapp.sql
@@ -3,7 +3,7 @@ CREATE TABLE ApiDB.PlasmoMapp (
  na_sequence_id  NUMERIC(10) NOT NULL,
  strand          NUMERIC(3)  NOT NULL,
  location        NUMERIC(10) NOT NULL,
- value	         FLOAT(10) NOT NULL,
+ value	         FLOAT NOT NULL,
  MODIFICATION_DATE     TIMESTAMP,
  USER_READ             NUMERIC(1),
  USER_WRITE            NUMERIC(1),
